WS-get firmware_id, Make log path based on firmware id
