+++ embark/embark/routing.py
@@ -6,5 +6,5 @@ from django.conf.urls import url
 # url patterns for websocket communication -> equivalent to urls.py
 # TODO: adjust url after progress example is removed
 ws_urlpatterns = [
-    path('ws/progress', consumers.WSConsumer.as_asgi()),
+    path('ws/progress/', consumers.WSConsumer.as_asgi()),
 ]
