+++ embark/templates/uploader/embaServiceDashboard.html
@@ -2,27 +2,9 @@
 {% block maincontent %}
 <div class="container-fluid sub-container">
     <div class="row">
-            <div class="coldiv">
-                    <a class="tile row statusTile">
-                            <!--<span class="headerText">Firmware Image</span>-->
-                            <div class="progress" id="progress-wrapper">
-                                    <div id="pBar" class="progress-bar" role="progressbar" style="width: 0%;" aria-valuenow="0" aria-valuemin="0" aria-valuemax="100">0%</div>
-                                    <br>
-                                    {% comment %} <div class="row uploadBtnRow">
-                                        <button type="submit" class="btn" onclick="embaProgress()">Progress</button>
-                                    </div> {% endcomment %}
-                            </div>
-
-                            <div class="row statusEMba">
-                                <div class="col-sm log tile moduleLog">
-                                        <ul class="log_phase" id="log_phase"> </ul>
-                                </div>
-                                <div class="col-sm log tile phaseLog">
-                                        <ul class="log_phase" id="log_module"> </ul>
-                                </div>
-                            </div>
-                    </a>
-            </div>
+        <a class="tile row statusTile">
+            <h1>i am a header</h1>
+        </a>
     </div>
 </div>
 {% endblock %}
