+++ embark/templates/uploader/fileUpload.html
@@ -93,7 +93,6 @@
                     {% endfor %}
 
                     <button type="submit" class="btn">Analyse</button>
-                    <button type="button" class="btn" onclick="embaProgress();">Show Progress</button>
                     <button type="reset" class="btn">Cancel</button>
 
                 </form>
